Moved MainPageActionHandler class to its own file.
